consolidated sfu_tasks and sfu_logging_queue to sfu_logging_task
